+++ src/serious_python/example/flet_example/app/src/main.py
@@ -1,9 +1,16 @@
 import logging
+import sys
+import urllib.request
 
 import flet as ft
 from flet_core.version import version
 
-logging.basicConfig(level=logging.DEBUG)
+# logging.basicConfig(level=logging.DEBUG)
+
+print("Hello from Python!")
+print(__name__)
+
+# import aaa
 
 
 def main(page: ft.Page):
@@ -13,12 +20,22 @@ def main(page: ft.Page):
     txt_number = ft.TextField(value="0", text_align=ft.TextAlign.RIGHT, width=100)
 
     def minus_click(e):
+        print("Clicked minus button")
         txt_number.value = str(int(txt_number.value) - 1)
         page.update()
 
     def plus_click(e):
         txt_number.value = str(int(txt_number.value) + 1)
         page.update()
+
+    def check_ssl(e):
+        try:
+            with urllib.request.urlopen("https://google.com") as res:
+                result = "OK"
+        except Exception as ex:
+            result = str(ex)
+        page.show_dialog(ft.AlertDialog(content=ft.Text(result)))
+        # print(result)
 
     page.add(
         ft.Row(
@@ -33,10 +50,18 @@ def main(page: ft.Page):
             expand=True,
         ),
         ft.Row(
-            [ft.Text(f"Flet version: {version}")],
+            [
+                ft.Text(f"Flet version: {version}"),
+                ft.OutlinedButton("Check SSL", on_click=check_ssl),
+                ft.OutlinedButton("Exit app", on_click=lambda _: sys.exit(100)),
+            ],
+            wrap=True,
             alignment=ft.MainAxisAlignment.CENTER,
         ),
     )
 
+    print("This is inside main() method!")
 
-ft.app(main)
+
+if __name__ == "__main__":
+    ft.app(main)