+++ src/serious_python_darwin/darwin/Classes/SeriousPythonPlugin.swift
@@ -32,6 +32,7 @@ public class SeriousPythonPlugin: NSObject, FlutterPlugin {
         case "runPython":
             let args: [String: Any] = call.arguments as? [String: Any] ?? [:]
             let appPath = args["appPath"] as! String
+            let script = args["script"] as? String
             let modulePaths = args["modulePaths"] as? [String] ?? []
             let envVars = args["environmentVariables"] as? [String:String] ?? [:]
             let sync = args["sync"] as? Bool ?? false
@@ -69,10 +70,19 @@ public class SeriousPythonPlugin: NSObject, FlutterPlugin {
             
             // run program either sync or in a thread
             if (sync) {
-                runPython(appPath: appPath)
+                if (script == nil) {
+                    runPythonFile(appPath: appPath)
+                } else {
+                    runPythonScript(script: script!)
+                }
             } else {
-                let t = Thread(target: self, selector: #selector(runPython), object: appPath)
-                t.start()
+                if (script == nil) {
+                    let t = Thread(target: self, selector: #selector(runPythonFile), object: appPath)
+                    t.start()
+                } else {
+                    let t = Thread(target: self, selector: #selector(runPythonScript), object: script!)
+                    t.start()
+                }
             }
             
             result(nil)
@@ -81,7 +91,7 @@ public class SeriousPythonPlugin: NSObject, FlutterPlugin {
         }
     }
     
-    @objc func runPython(appPath: String) {
+    @objc func runPythonFile(appPath: String) {
         Py_Initialize()
         
         // run app
@@ -89,6 +99,18 @@ public class SeriousPythonPlugin: NSObject, FlutterPlugin {
         let result = PyRun_SimpleFileEx(file, appPath, 1)
         if (result != 0) {
             print("Python program completed with error.")
+        }
+        
+        Py_Finalize()
+    }
+
+    @objc func runPythonScript(script: String) {
+        Py_Initialize()
+        
+        // run app
+        let result = PyRun_SimpleString(script)
+        if (result != 0) {
+            print("Python script completed with error.")
         }
         
         Py_Finalize()