+++ src/serious_python_linux/linux/serious_python_linux_plugin_private.h
@@ -9,5 +9,8 @@
 // Handles the getPlatformVersion method call.
 FlMethodResponse *get_platform_version();
 
-void run_python_script(gchar* appPath);
+void run_python_program(gchar *appPath);
+gpointer run_python_program_async(gpointer data);
+
+void run_python_script(gchar *script);
 gpointer run_python_script_async(gpointer data);