+++ src/serious_python_platform_interface/lib/serious_python_platform_interface.dart
@@ -30,7 +30,8 @@ abstract class SeriousPythonPlatform extends PlatformInterface {
   }
 
   Future<String?> run(String appPath,
-      {List<String>? modulePaths,
+      {String? script,
+      List<String>? modulePaths,
       Map<String, String>? environmentVariables,
       bool? sync}) {
     throw UnimplementedError('run() has not been implemented.');