+++ src/serious_python_platform_interface/lib/src/method_channel_serious_python.dart
@@ -16,11 +16,13 @@ class MethodChannelSeriousPython extends SeriousPythonPlatform {
 
   @override
   Future<String?> run(String appPath,
-      {List<String>? modulePaths,
+      {String? script,
+      List<String>? modulePaths,
       Map<String, String>? environmentVariables,
       bool? sync}) async {
     final Map<String, dynamic> arguments = {
       'appPath': appPath,
+      'script': script,
       'modulePaths': modulePaths,
       'environmentVariables': environmentVariables,
       'sync': sync