+++ src/serious_python_windows/lib/serious_python_windows.dart
@@ -23,12 +23,14 @@ class SeriousPythonWindows extends SeriousPythonPlatform {
 
   @override
   Future<String?> run(String appPath,
-      {List<String>? modulePaths,
+      {String? script,
+      List<String>? modulePaths,
       Map<String, String>? environmentVariables,
       bool? sync}) async {
     final Map<String, dynamic> arguments = {
       'exePath': Platform.resolvedExecutable,
       'appPath': appPath,
+      'script': script,
       'modulePaths': modulePaths,
       'environmentVariables': environmentVariables,
       'sync': sync